+++ src/main/java/dev/zakalren/pickmeup/cart/dto/UpdateCartItemRequest.java
@@ -6,6 +6,6 @@ import jakarta.validation.constraints.NotNull;
 public record UpdateCartItemRequest(
         @NotNull(message = "Quantity must not be null.")
         @Min(value = 1, message = "Quantity must be equal or more than 1.")
-        Long quantity
+        Integer quantity
 ) {
 }
